IoT/requirements.txt
ecdsa
fastecdsa
pycryptodome
prometheus_client